+++ sumo/src/gui/TraCIServerAPI_GUI.cpp
@@ -117,7 +117,7 @@ TraCIServerAPI_GUI::processGet(TraCIServer &server, tcpip::Storage &inputStorage
     server.writeStatusCmd(CMD_GET_GUI_VARIABLE, RTYPE_OK, warning, outputStorage);
     // send response
     outputStorage.writeUnsignedByte(0); // command length -> extended
-    outputStorage.writeInt(1 + 4 + tempMsg.size());
+    outputStorage.writeInt(static_cast<int>(1 + 4 + tempMsg.size()));
     outputStorage.writeStorage(tempMsg);
     return true;
 }
